requirements.txt
asgiref==3.3.1
Django==3.1.4
django-autoslug==1.9.8
django-cors-headers==3.5.0
djangorestframework==3.12.2
Pillow==8.0.1
pytz==2020.4
sqlparse==0.4.1
